api/requirements.txt
channels==4.0.0
channels-redis==4.0.0
daphne==4.0.0
Django==4.1.6
django-cors-headers==3.13.0
django-lifecycle==1.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
psycopg2-binary==2.9.5